requirements.txt


streamlit>=1.30
numpy>=1.23
opencv-python>=4.8
tensorflow>=2.12
pillow>=9.5
matplotlib>=3.7